refactor: API 구조 변경
- TYPE 파일 생성
- 각 API에 대한 파일 생성
- 기존 통합된 API 파일 삭제
