+++ src/api/studyRecord/types.ts
@@ -1,0 +1,18 @@
+export type StudyTimeResponse = {
+  todayStudyTime: string; // 예: "0:00:01"
+  totalStudyTime: string; // 예: "139:06:58"
+};
+
+export type AttendanceResponse = {
+  attendance: boolean;
+};
+
+export type UpdateStudyTimePayload = {
+  todayStudyTime: string;
+};
+
+export type ApiResponse<T> = {
+  success: boolean;
+  response: T;
+  error: any;
+};
